+++ src/test/java/com/github/dakusui/scriptiveunit/examples/DryQapi.java
@@ -3,8 +3,8 @@ package com.github.dakusui.scriptiveunit.examples;
 import com.github.dakusui.scriptiveunit.annotations.Load;
 import com.github.dakusui.scriptiveunit.core.Config;
 import com.github.dakusui.scriptiveunit.model.lang.HostSpec;
-import com.github.dakusui.scriptiveunit.loaders.json.JsonPreprocessorUtils;
-import com.github.dakusui.scriptiveunit.model.lang.json.JsonUtils;
+import com.github.dakusui.scriptiveunit.testutils.JsonPreprocessorUtils;
+import com.github.dakusui.scriptiveunit.utils.JsonUtils;
 import com.github.dakusui.scriptiveunit.model.lang.ApplicationSpec;
 import com.github.dakusui.scriptiveunit.unittests.core.UtJsonUtils;
 import com.github.dakusui.scriptiveunit.utils.ReflectionUtils;
@@ -18,7 +18,7 @@ import java.io.ByteArrayInputStream;
 import java.io.InputStream;
 import java.nio.charset.Charset;
 
-import static com.github.dakusui.scriptiveunit.loaders.json.JsonPreprocessorUtils.requireObjectNode;
+import static com.github.dakusui.scriptiveunit.utils.JsonUtils.requireObjectNode;
 
 @Load(with = DryQapi.Loader.class)
 public class DryQapi extends Qapi {
