+++ src/test/java/com/github/dakusui/scriptiveunit/examples/SuiteSetExample.java
@@ -1,7 +1,7 @@
 package com.github.dakusui.scriptiveunit.examples;
 
-import com.github.dakusui.scriptiveunit.utils.ScriptiveSuiteSet;
-import com.github.dakusui.scriptiveunit.utils.ScriptiveSuiteSet.SuiteScripts;
+import com.github.dakusui.scriptiveunit.runners.ScriptiveSuiteSet;
+import com.github.dakusui.scriptiveunit.runners.ScriptiveSuiteSet.SuiteScripts;
 import org.junit.runner.RunWith;
 
 @RunWith(ScriptiveSuiteSet.class)
