+++ src/test/java/com/github/dakusui/scriptiveunit/testassets/drivers/ExampleSuiteSet.java
@@ -1,8 +1,8 @@
 package com.github.dakusui.scriptiveunit.testassets.drivers;
 
 
-import com.github.dakusui.scriptiveunit.utils.ScriptiveSuiteSet;
-import com.github.dakusui.scriptiveunit.utils.ScriptiveSuiteSet.SuiteScripts;
+import com.github.dakusui.scriptiveunit.runners.ScriptiveSuiteSet;
+import com.github.dakusui.scriptiveunit.runners.ScriptiveSuiteSet.SuiteScripts;
 import com.github.dakusui.scriptiveunit.examples.Qapi;
 import org.junit.runner.RunWith;
