+++ src/test/java/com/github/dakusui/scriptiveunit/testutils/JsonPreprocessorUtils.java
@@ -1,4 +1,4 @@
-package com.github.dakusui.scriptiveunit.loaders.json;
+package com.github.dakusui.scriptiveunit.testutils;
 
 import org.codehaus.jackson.JsonNode;
 import org.codehaus.jackson.node.ArrayNode;
@@ -9,16 +9,11 @@ import org.codehaus.jackson.node.TextNode;
 import java.util.AbstractList;
 
 import static com.github.dakusui.scriptiveunit.exceptions.SyntaxException.nonArray;
-import static com.github.dakusui.scriptiveunit.exceptions.SyntaxException.nonObject;
 import static com.github.dakusui.scriptiveunit.exceptions.SyntaxException.nonText;
 import static com.github.dakusui.scriptiveunit.utils.Checks.check;
 
 public enum JsonPreprocessorUtils {
   ;
-
-  public static ObjectNode requireObjectNode(JsonNode curr) {
-    return (ObjectNode) check(curr, v -> curr.isObject(), () -> nonObject(curr));
-  }
 
   private static TextNode requireTextNode(JsonNode curr) {
     return (TextNode) check(curr, v -> curr.isTextual(), () -> nonText(curr));
