+++ src/test/java/com/github/dakusui/scriptiveunit/unittests/cli/SuiteSetDriver.java
@@ -1,6 +1,6 @@
 package com.github.dakusui.scriptiveunit.unittests.cli;
 
-import com.github.dakusui.scriptiveunit.utils.ScriptiveSuiteSet;
+import com.github.dakusui.scriptiveunit.runners.ScriptiveSuiteSet;
 import org.junit.runner.RunWith;
 
 @RunWith(ScriptiveSuiteSet.class)
